Fix build issues in base/ for Android x64
This CL fixes several issues when building files under base/ for
Android x64.

BUG=346626

Review URL: https://codereview.chromium.org/179373003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254124 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f690ed9164c29b905495632fa61fe17981fd7f51

+++ base/android/java_handler_thread.cc
@@ -46,7 +46,8 @@ void JavaHandlerThread::Start() {
 void JavaHandlerThread::Stop() {
 }
 
-void JavaHandlerThread::InitializeThread(JNIEnv* env, jobject obj, jint event) {
+void JavaHandlerThread::InitializeThread(JNIEnv* env, jobject obj,
+                                         jlong event) {
   // TYPE_JAVA to get the Android java style message loop.
   message_loop_.reset(new base::MessageLoop(base::MessageLoop::TYPE_JAVA));
   static_cast<MessageLoopForUI*>(message_loop_.get())->Start();
